Move contents of SiStripBackPlaneCorrectionDepESProducer header to source
Also remove pset_ member as unnecessary
